requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
absl-py==2.3.1
    # via tensorboard
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via
    #   ipython
    #   tqdm
comm==0.2.3
    # via ipykernel
debugpy==1.8.15
    # via ipykernel
decorator==5.2.1
    # via ipython
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
grpcio==1.74.0
    # via tensorboard
hf-xet==1.1.5
    # via huggingface-hub
huggingface-hub==0.33.4
    # via
    #   -r requirements.in
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
ipykernel==6.30.0
    # via -r requirements.in
ipython==9.4.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via scikit-learn
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
lightning-utilities==0.14.3
    # via
    #   pytorch-lightning
    #   torchmetrics
markdown==3.8.2
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torchmetrics
    #   transformers
packaging==25.0
    # via
    #   huggingface-hub
    #   ipykernel
    #   lightning-utilities
    #   pytorch-lightning
    #   tensorboard
    #   torchmetrics
    #   transformers
pandas==2.3.1
    # via -r requirements.in
parso==0.8.4
    # via jedi
pillow==11.3.0
    # via tensorboard
platformdirs==4.3.8
    # via jupyter-core
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via tensorboard
psutil==7.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytorch-lightning==2.5.2
    # via -r requirements.in
pytz==2025.2
    # via pandas
pywin32==311
    # via jupyter-core
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
pyzmq==27.0.1
    # via
    #   ipykernel
    #   jupyter-client
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.16.0
    # via scikit-learn
setuptools==80.9.0
    # via
    #   lightning-utilities
    #   tensorboard
    #   torch
six==1.17.0
    # via python-dateutil
stack-data==0.6.3
    # via ipython
sympy==1.14.0
    # via torch
tensorboard==2.20.0
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchmetrics
torchmetrics==1.7.4
    # via pytorch-lightning
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.53.2
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   huggingface-hub
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via tensorboard
yarl==1.20.1
    # via aiohttp
